this is going to be the gdb version of the mmix simulator on the virtual motherboard. Initialy it contains the basic files from the normal mmix simulator for the virtual motherboard.
It will be stripped from all interaction with the user.
A remote interface to gdb will be added.
A tracing facility will remain.
There will be no loading of application programms as
this will be done through gdb in the elf format.

A special implementation of gdb for the mmix architecture
is required but I hope this can be incorporated into
the regular gdb distribution.


git-svn-id: https://svn.code.sf.net/p/vmb/src/trunk@93 a22778cc-470f-459f-8348-c59b68ef04d0